+++ src/vector/arithmetic/vector_add_dd.cpp
@@ -1,4 +1,4 @@
-#include<DD-AVX.hpp>
+#include<DD-AVX_internal.hpp>
 
 using namespace ddavx_core;
 
@@ -16,7 +16,7 @@ void dd_real_vector::add(dd_real_vector& vec1, dd_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		size_t i=0, is=0, ie=0;
+		int i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
 			SIMDreg a_hi = load(hi[i]);
@@ -49,7 +49,7 @@ void dd_real_vector::add(d_real_vector& vec1, dd_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		size_t i=0, is=0, ie=0;
+		int i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
 			SIMDreg a_hi = load(hi[i]);
@@ -88,7 +88,7 @@ void dd_real_vector::add(d_real_vector& vec1, d_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		size_t i=0, is=0, ie=0;
+		int i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
 			SIMDreg a_hi = load(hi[i]);
