+++ src/vector/basic_op/vector_cast.cpp
@@ -1,10 +1,10 @@
-#include<DD-AVX.hpp>
+#include<DD-AVX_internal.hpp>
 
 d_real_vector:: operator dd_real_vector()
 {
 	printf("cast dd=>d\n");
 	dd_real_vector tmp;
-	for(size_t i=0; i<size(); i++){
+	for(int i=0; i<size(); i++){
 		tmp.hi[i]=*this[i];
 		tmp.lo[i]=0.0;
 	}
