+++ src/vector/blas/axpyz.cpp
@@ -1,4 +1,4 @@
-#include<DD-AVX.hpp>
+#include<DD-AVX_internal.hpp>
 using namespace ddavx_core;
 
 namespace dd_avx{
@@ -12,7 +12,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -47,7 +47,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -82,7 +82,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -117,7 +117,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -152,7 +152,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -187,7 +187,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -222,7 +222,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -257,7 +257,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -293,7 +293,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -327,7 +327,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -361,7 +361,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -395,7 +395,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -429,7 +429,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -463,7 +463,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -497,7 +497,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -531,7 +531,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
